@@ -1,32 +1,27 @@
 {
-  "_args": [
-    [
-      "normalize-url@4.5.0",
-      "/Users/Rabee/Desktop/meetme/meetme"
-    ]
-  ],
-  "_development": true,
-  "_from": "normalize-url@4.5.0",
-  "_id": "normalize-url@4.5.0",
+  "_from": "normalize-url",
+  "_id": "normalize-url@5.0.0",
   "_inBundle": false,
-  "_integrity": "sha512-2s47yzUxdexf1OhyRi4Em83iQk0aPvwTddtFz4hnSSw9dCEsLEGf6SwIO8ss/19S9iBb5sJaOuTvTGDeZI00BQ==",
+  "_integrity": "sha512-bAEm2fx8Dq/a35Z6PIRkkBBJvR56BbEJvhpNtvCZ4W9FyORSna77fn+xtYFjqk5JpBS+fMnAOG/wFgkQBmB7hw==",
   "_location": "/normalize-url",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "tag",
     "registry": true,
-    "raw": "normalize-url@4.5.0",
+    "raw": "normalize-url",
     "name": "normalize-url",
     "escapedName": "normalize-url",
-    "rawSpec": "4.5.0",
+    "rawSpec": "",
     "saveSpec": null,
-    "fetchSpec": "4.5.0"
+    "fetchSpec": "latest"
   },
   "_requiredBy": [
-    "/cacheable-request"
+    "#USER",
+    "/"
   ],
-  "_resolved": "https://registry.npmjs.org/normalize-url/-/normalize-url-4.5.0.tgz",
-  "_spec": "4.5.0",
+  "_resolved": "https://registry.npmjs.org/normalize-url/-/normalize-url-5.0.0.tgz",
+  "_shasum": "f46c9dc20670495e4e18fbd1b4396e41d199f63c",
+  "_spec": "normalize-url",
   "_where": "/Users/Rabee/Desktop/meetme/meetme",
   "author": {
     "name": "Sindre Sorhus",
@@ -36,21 +31,24 @@
   "bugs": {
     "url": "https://github.com/sindresorhus/normalize-url/issues"
   },
+  "bundleDependencies": false,
+  "deprecated": false,
   "description": "Normalize a URL",
   "devDependencies": {
     "ava": "^2.4.0",
     "coveralls": "^3.0.6",
-    "nyc": "^14.1.1",
-    "tsd": "^0.8.0",
-    "xo": "^0.24.0"
+    "nyc": "^15.0.0",
+    "tsd": "^0.11.0",
+    "xo": "^0.25.3"
   },
   "engines": {
-    "node": ">=8"
+    "node": ">=10"
   },
   "files": [
     "index.js",
     "index.d.ts"
   ],
+  "funding": "https://github.com/sponsors/sindresorhus",
   "homepage": "https://github.com/sindresorhus/normalize-url#readme",
   "keywords": [
     "normalize",
@@ -76,5 +74,5 @@
   "scripts": {
     "test": "xo && nyc ava && tsd"
   },
-  "version": "4.5.0"
+  "version": "5.0.0"
 }
