@@ -1,9 +1,23 @@
-import React from "react";
+import React, { useContext } from "react";
+import { Route, Redirect } from "react-router-dom";
+import authContext from "context/authContext/authContext";
 
-interface Props {}
 
-const PrivateRoute = (props: Props) => {
-  return <div></div>;
+const PrivateRoute = ({ component: Component, ...rest }) => {
+  const auth = useContext(authContext);
+  const { isAuthenticated, loading } = auth;
+  return (
+    <Route
+      {...rest}
+      render={(props) =>
+        auth && !isAuthenticated && loading ? (
+          <Redirect to="/" />
+        ) : (
+          <Component {...props} />
+        )
+      }
+    />
+  );
 };
 
 export default PrivateRoute;
