@@ -1,5 +1,5 @@
 import { createContext } from "react";
 
-const authContext = createContext({});
+const authContext = createContext<any | undefined>(undefined);
 
 export default authContext;
