@@ -1,22 +1,30 @@
-import * as express from "express";
 import jwt from "jsonwebtoken";
-import { Request, Response } from "express";
+import { Request, Response, NextFunction } from "express";
 import dotenv from "dotenv";
 
 dotenv.config();
 
-const secretJWT = process.env.jsonwebtoken!;
-// @Todo Ali change the auth function parameter default type with typescript
-export default function (req: any, res: any, next: express.NextFunction) {
-  const token = req.header("x-auth-token");
+export interface IPayload {
+  id: string;
+}
+
+const secretJWT: string = process.env.jwtSecret!;
+
+export default function (
+  req: Request,
+  res: Response,
+  next: NextFunction
+): Response<any> | undefined {
+  const token: string | undefined = req.header("x-auth-token");
 
   if (!token) {
     return res.status(401).json({ errors: [{ msg: "No token" }] });
   }
 
   try {
-    const decoded: any = jwt.verify(token, secretJWT);
-    req.user = decoded.user;
+    const payload = jwt.verify(token, secretJWT) as IPayload;
+    req.userId = payload.id;
+
     next();
   } catch (err) {
     res.status(401).json({ errors: [{ msg: "Token is not valid" }] });
